+++ src/routes/ConsortiumManager/settings/data-export/DataExportLogs/DataExportLogs.test.js
@@ -27,6 +27,11 @@ jest.mock('@folio/stripes-data-transfer-components', () => ({
   useJobLogsListFormatter: jest.fn(),
 }));
 
+jest.mock('../utils', ()=>({
+  ...jest.requireActual('../utils'),
+  getExportJobLogsListResultsFormatter: jest.fn(),
+}))
+
 jest.mock('../hooks', () => ({
   ...jest.requireActual('../hooks'),
   useDataExportLogs: jest.fn(),