+++ CHANGELOG.md
@@ -27,3 +27,4 @@
 * [UICONSET-138](https://issues.folio.org/browse/UICONSET-138) Upgrade `React` to `v18`
 * [UICONSET-137](https://issues.folio.org/browse/UICONSET-137) Update `Node.js` to `v18`s in GitHub Actions.
 * [UICONSET-112](https://issues.folio.org/browse/UICONSET-112) Allow a user to delete shared settings in the `<ConsortiaControlledVocabulary>`.
+* [UICONSET-139](https://issues.folio.org/browse/UICONSET-139) Consortia-settings: format numbers in "totalRecords" columns.